Fix: Triage IE bugs
There were two bugs I found from IE on the home page. First, it was having a hard time with both repeat.for and if.bind on the same element. I hadn't seen this before, but it makes sense that this would confuse IE. Instead, I've achieved the same behavior by using the filter value converter on the repeat.for.

Second, the above strategy threw an error in IE since String.prototype.includes is undefined in IE11. Since this is likely an issue anywhere that the filter value converter is used, I added a new "polyfills" feature and added a string.includes polyfill, taken from the MDN website.