Fix: Replace the new URL() function
The URL API is not supported in IE11, so I have removed its usage throughout the application. I removed its explicit usage in member-detail.ts, and replaced it with the dom_utils function this_page_url(). I rewrote the this_page_url() function to leverage the location API instead, which is widely supported.