Feature: Add a "tap and hold" custom element
After looking through the event and interact documentation, I didn't see a good ready-made solution to this problem. The mobile browsers tend to implement a long press behavior to open in a new tab, similar to ctrl+click on a desktop browser, but these are browser internals and not specific events. I've enabled this behavior through the "tapandhold" custom attribute. Adding this attribute will fire two new events, "longtouch" and "longclick". "longtouch" fires when the user taps and holds for x milliseconds. "longclick" fires when the user taps and holds for at least x milliseconds before releasing. The default value for x is 300 milliseconds, but you can customize this by passing a value to the custom attribute, e.g. tapandhold="500". I've added handlers on photos.html to demonstrate.