feature(app): Scale up menu fonts
To do this, I had to cut elsewhere. I shrank the brand font size a bit, shrunk the loader spinner, and removed some of the spacing between the menu items.

I also change the way mobile scales the menu fonts, using vw instead of a fixed font. This looks really good on phone devices. It gets a bit too big on the tablet portrait layout, but its not bad.